requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.4
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.6.0
Flask==2.2.3
Flask-Cors==3.0.10
frozenlist==1.3.3
future==0.18.3
google-api-core==2.11.0
google-auth==2.16.3
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.59.0
huggingface-hub==0.12.1
idna==3.4
importlib-metadata==5.0.0
importlib-resources==5.10.0
imutils==0.5.4
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.0
jupyter-server==1.23.2
jupyter_client==7.4.7
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mistune==2.0.4
more-itertools==9.1.0
multidict==6.0.4
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
openai==0.27.2
openai-whisper==20230124
opencv-python==4.7.0.72
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
platformdirs==2.5.2
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==4.22.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
PyYAML==6.0
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.2
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
slack-bolt==1.13.1
slack-sdk==3.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stable-ts @ git+https://github.com/jianfch/stable-ts.git@15792c74c9465ca8ab5b8ae84210df195725d542
stack-data==0.6.1
terminado==0.17.0
tinycss2==1.2.1
tokenizers==0.13.2
torch==1.13.1
torchaudio==0.13.1
tornado==6.2
tqdm==4.65.0
traitlets==5.5.0
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.14
virtualenv==20.14.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.10.0
